requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.13.2



